dom/webauthn/tests/test_webauthn_isplatformauthenticatoravailable.html

<!DOCTYPE html>
<meta charset=utf-8>
<head>
  <title>Test for W3C Web Authentication isPlatformAuthenticatorAvailable</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="u2futil.js"></script>
  <script type="text/javascript" src="pkijs/common.js"></script>
  <script type="text/javascript" src="pkijs/asn1.js"></script>
  <script type="text/javascript" src="pkijs/x509_schema.js"></script>
  <script type="text/javascript" src="pkijs/x509_simpl.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>

<h1>Test for W3C Web Authentication isPlatformAuthenticatorAvailable</h1>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1309284">Mozilla Bug 1309284</a>

<script class="testbody" type="text/javascript">
"use strict";

// Execute the full-scope test
SimpleTest.waitForExplicitFinish();

// Turn off all tokens. This should result in "not allowed" failures
SpecialPowers.pushPrefEnv({"set": [["security.webauth.webauthn", true],
                                   ["security.webauth.webauthn_enable_softtoken", true],
                                   ["security.webauth.webauthn_enable_usbtoken", false]]},
function() {
  PublicKeyCredential.isPlatformAuthenticatorAvailable()
  .then(function(aResult) {
    // The specification requires this method, if will return false, to wait 10
    // minutes for anti-fingerprinting reasons. So we really can't test that
    // in an automated way.
    ok(aResult, "Should be available!");
  })
  .catch(function(aProblem) {
    is(false, "Problem encountered: " + aProblem);
  })
  .then(function() {
    SimpleTest.finish();
  })
});

</script>

</body>
</html>
